Reuse shared readLastLines for PHP error tail instead of a custom helper
